Updated the GitHub workflow to handle SDK and design system version updates during the release process.
